Removes version-based database updates
Removes the logic that updated the database based on app version changes.  The database will now be updated based on a configurable cache duration or when manually triggered.

This simplifies the update process and removes the dependency on the `package_info_plus` package.  A local boolean constant `_forceRefresh` has been added to allow for manual cleaning of the database during development.